Upgraded from 16.04 to 20.04
Took the opportunity to create one RUN command, which creates much
smaller images.

When you update Ubuntu 18.04 and later on Docker, you'll get a prompt
asking you what timezone you're in.  This doesn't work under Docker,
thus we need to install tzdata and set various settings to shut it up.

This should have also pulled the latest stable DokuWiki -- as of the
time of writing, there is a release candidate that's not getting pulled
in, will guess shortly it'll become stable and will get pulled in.